[WINDOWS] Use timer instead of onIdle to run GUI.
When you open a menu on Windows, the GUI stops running until it is
closed. This also applies to menu items mapped to keyboard, such as save
state. Due to this, there is a micro pause whenever a shortcut is run,
that interrupts input being pressed.

With this design, the app should execute its main loop according to a
timer. It is no longer attached to any window modal, at least not
explicitly.

Since this is only necessary on Windows, we protect all changes
with `__WXMSW__`.

- Fix #83.
- Fix #587.